Add WoWSpec function get_wow_spec_from_combined_name
This adds a easy way to get a WoWSpec from the strings used in Simc automated CI tests which are of the form
Class_Spec with white space in Class being tokenized with underscores.

Since eg. Death_Knight_Blood and Hunter_Beast_Mastery make it hard to handle this on the caller site, it was easier to add a special function here.